TP-RK-2 END stable (#30)
* TP update: migrations

* TP update: migrations

* TP update: migrations

* Update cors

* Update cors

* Update cors

* Update corsg

* Update corsg

* initial tests

* update tests

* fix lint

* Fix cookie

* fix images

* fix csrf

* TP update: rateFilm handler

* TP update: reviews handler

* TP update: signup handler

* TP update: time and date forman

* TP update: sql filler

* TP update: sql doc

* fix put image

* fix lint

* Add check filetype

* Fix auth repo errors, csrf response, refactoring

* TP update: time and date forman

* TP update: fix reviews count

* fix put image png

+++ internal/models/user.go
@@ -25,7 +25,7 @@ type NodeInUserCollection struct {
 
 type UserActivity struct {
 	CountReviews int                    `json:"count_reviews,omitempty"`
-	Rating       float32                `json:"rating,omitempty"`
+	Rating       int                    `json:"rating,omitempty"`
 	DateRating   string                 `json:"date_rating,omitempty"`
 	Collections  []NodeInUserCollection `json:"collections,omitempty"`
 }
